Use byte-string headers in benchmark
Obviously the benchmark is a bit harder if we use unicode strings
instead of byte-strings, but harder in a boring way: there's no changes
we can make to h11 to make str.encode faster, so better if we skip that
and focus the benchmark on things we actually can change.

This also makes the results more comparable between py2 and py3.