Adopt black and isort to autoformat the files
This removes any need to figure out how to style the code for h11,
instead allowing these two commands to do it instead,

    black h11/
    isort --dont-skip __init__.py --apply --settings-path setup.cfg --recursive h11

In addition a format stage has been added to the tox setup to enforce
this styling.

This is the same styling as used by the wsproto project.